requirements.txt
Flask==2.0.2
Flask_SQLAlchemy==2.5.1
PyJWT==2.3.0
python-dotenv==0.19.1
